Add configuration of Routes
