+++ src/components/Currency.jsx
@@ -1,10 +1,15 @@
 import PropTypes from 'prop-types';
+import { NavLink } from 'react-router-dom';
 
 const CurrencyComponent = ( { title, description, price } ) => (
   <div className="currency-data">
-    <h3>{title}</h3>
-    <p>{description}</p>
-    <p>{price}</p>
+    <NavLink
+      to={ title}
+    >
+      <h3>{title}</h3>
+      <p>{description}</p>
+      <p>{price}</p>
+   </NavLink>
   </div>
 )
